Delete review route checks user id on cookie matches review author id before deleting

+++ src/routes/reviews.js
@@ -20,7 +20,12 @@ router.post('/albums/:albumId/reviews/new', (req, res) => {
 })
 
 router.delete('/deletereview/:reviewId', (req) => {
-  reviews.deleteById(req.params.reviewId)
+  reviews.getById(req.params.reviewId)
+    .then((review) => {
+      if (review.user_id === req.session.user.id) {
+        reviews.deleteById(req.params.reviewId)
+      } else (console.error('Error deleting review'))
+    })
 })
 
 module.exports = router
